% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifo-package.R
\docType{package}
\name{ifo-package}
\alias{ifo}
\alias{ifo-package}
\title{ifo: Client for the Ifo Institute Time Series}
\description{
Download ifo business survey data and more time series from ifo institute \url{https://www.ifo.de/en/ifo-time-series}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://m-muecke.github.io/ifo/}
  \item \url{https://github.com/m-muecke/ifo}
  \item Report bugs at \url{https://github.com/m-muecke/ifo/issues}
}

}
\author{
\strong{Maintainer}: Maximilian Muecke \email{muecke.maximilian@gmail.com} (\href{https://orcid.org/0009-0000-9432-9795}{ORCID})

}
\keyword{internal}
