% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifo.R
\name{ifo_empl}
\alias{ifo_empl}
\title{Return ifo employment expectations}
\usage{
ifo_empl()
}
\description{
Long time-series of the ifo Employment Barometer for Germany
}
\seealso{
Other ifo time series: 
\code{\link{ifo_climate}()},
\code{\link{ifo_eastern}()},
\code{\link{ifo_export}()},
\code{\link{ifo_saxony}()}
}
\concept{ifo time series}
