% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifo.R
\name{ifo_employment}
\alias{ifo_employment}
\title{Return ifo employment expectations}
\usage{
ifo_employment()
}
\value{
A \code{data.frame()} containing the ifo employment expectations.
}
\description{
Long time-series of the ifo Employment Barometer for Germany.
}
\examples{
\donttest{
ifo_employment()
}
}
\references{
\url{https://www.ifo.de/en/ifo-time-series}
}
\seealso{
Other ifo business survey: 
\code{\link{ifo_climate}()},
\code{\link{ifo_export}()}
}
\concept{ifo business survey}
