% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifo.R
\name{ifo_expectation}
\alias{ifo_expectation}
\title{Return ifo expectation data}
\source{
\url{https://www.ifo.de/en/ifo-time-series}
}
\usage{
ifo_expectation(type = c("export", "employment"))
}
\arguments{
\item{type}{(\code{character(1)})\cr
Defaults to \code{"export"}. One of:
\itemize{
\item \code{"export"}: returns the ifo export expectations for manufacturing.
\item \code{"employment"}: returns the ifo employment barometer for Germany.
}}
}
\value{
A \code{data.frame()} containing the monthly ifo expectation time series.
}
\description{
Return ifo expectation data
}
\examples{
\dontshow{if (curl::has_internet()) withAutoprint(\{ # examplesIf}
\donttest{
expectation <- ifo_expectation("export")
head(expectation)
}
\dontshow{\}) # examplesIf}
}
