% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifo.R
\name{ifo_export}
\alias{ifo_export}
\title{Return ifo export expectations}
\usage{
ifo_export()
}
\value{
A \code{data.frame()} containing the ifo export expectations.
}
\description{
Long time-series of the ifo Export Expectations for manufacturing.
}
\examples{
\donttest{
ifo_export()
}
}
\references{
\url{https://www.ifo.de/en/ifo-time-series}
}
\seealso{
Other ifo business survey: 
\code{\link{ifo_climate}()},
\code{\link{ifo_employment}()}
}
\concept{ifo business survey}
