% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifo.R
\name{ifo_export_climate}
\alias{ifo_export_climate}
\title{Return ifo export climate}
\usage{
ifo_export_climate()
}
\value{
A \code{data.frame()} containing the ifo export climate.
}
\description{
Long time series for the ifo Export Climate and the annual rate of change of real
exports.
}
\examples{
\donttest{
ifo_export_climate()
}
}
\references{
\url{https://www.ifo.de/en/ifo-time-series}
}
\seealso{
Other ifo time series: 
\code{\link{ifo_import_climate}()}
}
\concept{ifo time series}
