% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifo.R
\name{ifo_import_climate}
\alias{ifo_import_climate}
\title{Return ifo import climate}
\usage{
ifo_import_climate()
}
\value{
A \code{data.frame()} containing the ifo import climate.
}
\description{
Long time-series of the ifo import climate.
}
\examples{
\donttest{
ifo_import_climate()
}
}
\references{
Grimme C, Lehmann R, Nöller M (2018).
\dQuote{Das ifo Importklima -- ein erster Frühindikator für die Prognose der deutschen Importe.}
\emph{ifo Schnelldienst}, \bold{71}(12), 27--32.

Grimme, Christian, Lehmann, Robert, Nöller, Marvin (2021).
\dQuote{Forecasting imports with information from abroad.}
\emph{Economic Modelling}, \bold{98}, 109--117.
}
\seealso{
Other ifo time series: 
\code{\link{ifo_export_climate}()}
}
\concept{ifo time series}
