% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifo.R
\name{ifo_sachsen}
\alias{ifo_sachsen}
\title{Return ifo business climate for Saxony}
\usage{
ifo_sachsen()
}
\description{
Long time-series of the Business Climate for Saxony and its two components,
the business situation and the expectations.
}
\references{
\url{https://www.ifo.de/en/ifo-time-series}
}
\seealso{
Other ifo time series: 
\code{\link{ifo_empl}()},
\code{\link{ifo_export}()},
\code{\link{ifo_ostd}()}
}
\concept{ifo time series}
