% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifo.R
\name{ifo_saxony}
\alias{ifo_saxony}
\title{Return ifo business climate for Saxony}
\usage{
ifo_saxony()
}
\description{
Long time-series of the Business Climate for Saxony and its two components,
the business situation and the expectations.
}
\seealso{
Other ifo time series: 
\code{\link{ifo_climate}()},
\code{\link{ifo_eastern}()},
\code{\link{ifo_empl}()},
\code{\link{ifo_export}()}
}
\concept{ifo time series}
