% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifo.R
\name{ifo_world_climate}
\alias{ifo_world_climate}
\title{Return ifo World Economic Climate data}
\usage{
ifo_world_climate(type = c("world", "euro"))
}
\arguments{
\item{type}{\code{character(1)} Defaults to \code{"world"}. One of:
\itemize{
\item \code{"world"}: returns the ifo world economic climate.
\item \code{"euro"}: returns the ifo world economic climate for the euro zone.
}}
}
\value{
A \code{data.frame()} containing the quarterly ifo climate time series.
}
\description{
Return ifo World Economic Climate data
}
\examples{
\dontrun{
world <- ifo_world_climate("world")
euro <- ifo_world_climate("euro")
}
}
\references{
\url{https://www.ifo.de/en/ifo-time-series}
}
