% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifo.R
\name{ifo_business}
\alias{ifo_business}
\title{Return ifo business climate}
\usage{
ifo_business(
  type = c("climate", "sectors", "eastern", "saxony"),
  long_format = TRUE
)
}
\arguments{
\item{type}{\code{character(1)} one of \code{"climate"}, \code{"sectors"},\code{"eastern"},
\code{"saxony"}. Default \code{"climate"}.}

\item{long_format}{\code{logical(1)} if \code{TRUE} return the data in long format.
Only applies to \code{type} \code{"climate"} and \code{"sectors"}. Default \code{TRUE}.}
}
\value{
A \code{data.frame()} containing the monthly ifo business climate time series.
}
\description{
Long time-series of the ifo Business Climate for Germany and its two and
its two components, the business situation, and the expectations for sectors.
}
\examples{
\donttest{
ifo_business()
}
}
\references{
\url{https://www.ifo.de/en/ifo-time-series}
}
\concept{ifo business survey}
