% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifo.R
\name{ifo_climate}
\alias{ifo_climate}
\title{Return ifo business climate}
\usage{
ifo_climate(type = c("climate", "sectors", "eastern", "saxony"))
}
\arguments{
\item{type}{\code{character(1)} one of \code{"germany"}, \code{"eastern"}, \code{"saxony"}.
Default \code{"germany"}.}
}
\description{
Long time-series of the ifo Business Climate for Germany and its two and
its two components, the business situation, and the expectations for sectors.
}
\examples{
ifo_climate()
}
\references{
\url{https://www.ifo.de/en/ifo-time-series}
}
\seealso{
Other ifo time series: 
\code{\link{ifo_employment}()},
\code{\link{ifo_export}()}
}
\concept{ifo time series}
