% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifo.R
\name{ifo_employment}
\alias{ifo_employment}
\title{Return ifo employment expectations}
\usage{
ifo_employment()
}
\description{
Long time-series of the ifo Employment Barometer for Germany.
}
\examples{
\donttest{
ifo_employment()
}
}
\seealso{
Other ifo time series: 
\code{\link{ifo_climate}()},
\code{\link{ifo_export}()}
}
\concept{ifo time series}
