% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifo.R
\name{ifo_expectation}
\alias{ifo_expectation}
\title{Return ifo expectation data}
\usage{
ifo_expectation(type = c("export", "employment"))
}
\arguments{
\item{type}{\code{character(1)} Defaults to \code{"employment"}. One of:
\itemize{
\item \code{"export"}: returns the ifo export expectations for manufacturing.
\item \code{"employment"}: returns the ifo employment barometer for Germany.
}}
}
\value{
A \code{data.frame()} containing the monthly ifo expectation time series.
}
\description{
Return ifo expectation data
}
\examples{
\dontrun{
export <- ifo_expectation("export")
employment <- ifo_expectation("emplpoyment")
}
}
\references{
\url{https://www.ifo.de/en/ifo-time-series}
}
