% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifo.R
\name{ifo_export}
\alias{ifo_export}
\title{Return ifo export expectations}
\usage{
ifo_export()
}
\description{
Long time-series of the ifo Export Expectations for manufacturing
}
\references{
\url{https://www.ifo.de/en/ifo-time-series}
}
\seealso{
Other ifo time series: 
\code{\link{ifo_empl}()},
\code{\link{ifo_ostd}()},
\code{\link{ifo_sachsen}()}
}
\concept{ifo time series}
