% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifo.R
\name{ifo_export}
\alias{ifo_export}
\title{Return ifo export expectations}
\usage{
ifo_export()
}
\description{
Long time-series of the ifo Export Expectations for manufacturing
}
\seealso{
Other ifo time series: 
\code{\link{ifo_climate}()},
\code{\link{ifo_eastern}()},
\code{\link{ifo_empl}()},
\code{\link{ifo_saxony}()}
}
\concept{ifo time series}
