% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifo.R
\name{ifo_export}
\alias{ifo_export}
\title{Return ifo export expectations}
\usage{
ifo_export()
}
\description{
Long time-series of the ifo Export Expectations for manufacturing.
}
\examples{
ifo_export()
}
\seealso{
Other ifo time series: 
\code{\link{ifo_climate}()},
\code{\link{ifo_employment}()}
}
\concept{ifo time series}
