% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifo.R
\name{ifo_ostd}
\alias{ifo_ostd}
\title{Return ifo business climate for eastern Germany}
\usage{
ifo_ostd()
}
\description{
Long time-series of the Business Climate for Eastern Germany and its two
components, the business situation and the expectations.
}
\references{
\url{https://www.ifo.de/en/ifo-time-series}
}
